# docs/openapi.yaml
openapi: "3.0.3"
info:
  title: Weave API
  description: Chat-based multi-bank money transfer agent API
  version: "1.0.0"
servers:
  - url: https://api.weave.io
    description: Production
  - url: http://localhost:8080
    description: Local development
paths:
...
  /api/v1/banks/link:
    post:
      summary: Initiate bank account linking
      operationId: initiateLink
      tags: [Banks]
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [provider]
              properties:
                provider:
                  type: string
                  enum: [okra, mono]
      responses:
        "200":
          description: Link URL generated
          content:
            application/json:
              schema:
                type: object
                properties:
                  connect_url:
                    type: string
                  reference:
                    type: string
                  provider:
                    type: string
        "503":
          $ref: "#/components/responses/APIError"
  /api/v1/banks/webhook/okra:
    post:
      summary: Okra account-linking webhook
      operationId: okraWebhook
      tags: [Banks]
      responses:
        "200":
          description: Account linked
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
        "401":
          $ref: "#/components/responses/APIError"
  /api/v1/banks/webhook/mono:
    post:
      summary: Mono account-linking webhook
      operationId: monoWebhook
      tags: [Banks]
      responses:
        "200":
          description: Account linked
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
        "401":
          $ref: "#/components/responses/APIError"
  /api/v1/banks:
    get:
      summary: List linked bank accounts
      operationId: listBanks
      tags: [Banks]
      security: [{ BearerAuth: [] }]
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        "200":
          description: Linked bank accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BankAccount"
  /api/v1/banks/{id}/priority:
    put:
      summary: Update bank account priority
      operationId: updatePriority
      tags: [Banks]
      security: [{ BearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [priority]
              properties:
                priority:
                  type: integer
                  minimum: 1
                  maximum: 5
      responses:
        "200":
          description: Priority updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: updated
  /api/v1/banks/{id}:
    delete:
      summary: Unlink a bank account
      operationId: unlinkBank
      tags: [Banks]
      security: [{ BearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Account unlinked
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: deleted
  /api/v1/banks/{id}/refresh:
    post:
      summary: Refresh bank account balance from provider
      operationId: refreshBalance
      tags: [Banks]
      security: [{ BearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Balance refreshed
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: refreshed
                  last_balance:
                    type: integer
                    description: Balance in kobo
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    APIError:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
...
    BankAccount:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        provider:
          type: string
          enum: [okra, mono]
        account_number:
          type: string
        account_name:
          type: string
        bank_code:
          type: string
        bank_name:
          type: string
        priority:
          type: integer
        min_balance:
          type: integer
          description: Cushion balance in kobo
        last_balance:
          type: integer
          description: Last known balance in kobo
        is_active:
          type: boolean
        is_verified:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
  responses:
    APIError:
      description: Error response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/APIError"